Add host attribute to the Agent object
For a local project, we found that the host attribute is being returned from the
TeamCity server, but we are unable to access it in the Agent object.  According
to the online docs, this is an official attribute so likely it was added since
the last time the swagger.json was pulled.

Ideally we would pull the new swagger.json and re-generate everything based on
the new swagger.json, but I found an enormous number of differences, including
breaking changes if we do that.  For now, just adding this attribute is
sufficient to make forward progress